backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
psycopg2-binary
redis
aioredis
python-jose[cryptography]
bcrypt
python-multipart
PyJWT
pydantic[email]
pydantic-settings
xgboost
scikit-learn
numpy
pandas
httpx
python-dotenv
